# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "DriverProfileOut": {
        "properties": {
          "escrow_balance": {
            "title": "Escrow Balance",
            "type": "string"
          },
          "escrow_status": {
            "$ref": "#/components/schemas/EscrowStatus"
          },
          "escrow_target": {
            "title": "Escrow Target",
            "type": "string"
          },
          "fraud_strikes": {
            "title": "Fraud Strikes",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "license_number": {
            "title": "License Number",
            "type": "string"
          },
          "license_years_experience": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Years Experience"
          },
          "online_status": {
            "$ref": "#/components/schemas/OnlineStatus"
          },
          "rating_avg": {
            "title": "Rating Avg",
            "type": "string"
          },
          "total_trips": {
            "title": "Total Trips",
            "type": "integer"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "license_number",
          "license_years_experience",
          "online_status",
          "escrow_balance",
          "escrow_target",
          "escrow_status",
          "rating_avg",
          "total_trips",
          "fraud_strikes"
        ],
        "title": "DriverProfileOut",
        "type": "object"
      },
      "EkycResult": {
        "properties": {
          "national_id_masked": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "National Id Masked"
          },
          "national_id_verified": {
            "title": "National Id Verified",
            "type": "boolean"
          }
        },
        "required": [
          "national_id_verified",
          "national_id_masked"
        ],
        "title": "EkycResult",
        "type": "object"
      },
      "EkycSubmit": {
        "properties": {
          "national_id_number": {
            "maxLength": 20,
            "minLength": 9,
            "title": "National Id Number",
            "type": "string"
          },
          "selfie_reference_url": {
            "title": "Selfie Reference Url",
            "type": "string"
          }
        },
        "required": [
          "national_id_number",
          "selfie_reference_url"
        ],
        "title": "EkycSubmit",
        "type": "object"
      },
      "EscrowStatus": {
        "enum": [
          "accumulating",
          "fulfilled"
        ],
        "title": "EscrowStatus",
        "type": "string"
      },
...
      "GoOnlineResponse": {
        "properties": {
          "online_status": {
            "$ref": "#/components/schemas/OnlineStatus"
          },
          "qr_token": {
            "title": "Qr Token",
            "type": "string"
          }
        },
        "required": [
          "online_status",
          "qr_token"
        ],
        "title": "GoOnlineResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LocationUpdate": {
        "properties": {
          "lat": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat",
            "type": "number"
          },
          "lng": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lng",
            "type": "number"
          }
        },
        "required": [
          "lat",
          "lng"
        ],
        "title": "LocationUpdate",
        "type": "object"
      },
      "OnlineStatus": {
        "enum": [
          "offline",
          "online",
          "on_trip"
        ],
        "title": "OnlineStatus",
        "type": "string"
      },
...
      "SelfieCheckResult": {
        "properties": {
          "account_status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "match_score": {
            "title": "Match Score",
            "type": "number"
          },
          "passed": {
            "title": "Passed",
            "type": "boolean"
          }
        },
        "required": [
          "passed",
          "match_score",
          "account_status"
        ],
        "title": "SelfieCheckResult",
        "type": "object"
      },
      "SelfieCheckSubmit": {
        "properties": {
          "selfie_url": {
            "title": "Selfie Url",
            "type": "string"
          }
        },
        "required": [
          "selfie_url"
        ],
        "title": "SelfieCheckSubmit",
        "type": "object"
      },
...
      "UserOut": {
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "national_id_masked": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "National Id Masked"
          },
          "national_id_verified": {
            "title": "National Id Verified",
            "type": "boolean"
          },
          "phone": {
            "title": "Phone",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          }
        },
        "required": [
          "id",
          "phone",
          "full_name",
          "role",
          "status",
          "national_id_verified"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "UserRole": {
        "enum": [
          "rider",
          "driver",
          "admin"
        ],
        "title": "UserRole",
        "type": "string"
      },
      "UserStatus": {
        "enum": [
          "active",
          "suspended",
          "banned"
        ],
        "title": "UserStatus",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Dán access_token từ /api/v1/auth/verify-otp",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "GoAn Backend",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/drivers/me": {
      "get": {
        "operationId": "my_driver_profile_api_v1_drivers_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverProfileOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "My Driver Profile",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/drivers/me/ekyc": {
      "post": {
        "operationId": "submit_ekyc_api_v1_drivers_me_ekyc_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EkycSubmit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EkycResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit Ekyc",
        "tags": [
          "users"
        ]
      }
    },
...
    "/api/v1/drivers/me/location": {
      "post": {
        "operationId": "update_location_api_v1_drivers_me_location_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Location",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/drivers/me/offline": {
      "post": {
        "operationId": "go_offline_api_v1_drivers_me_offline_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Go Offline",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/drivers/me/online": {
      "post": {
        "description": "Bật nhận chuyến: sinh QR động mới cho phiên online này (SPEC 7.1).",
        "operationId": "go_online_api_v1_drivers_me_online_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoOnlineResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Go Online",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/drivers/me/selfie-check": {
      "post": {
        "description": "Selfie ngẫu nhiên chống tráo tài xế; không khớp -> khoá tài khoản ngay (SPEC 7.4).",
        "operationId": "selfie_check_api_v1_drivers_me_selfie_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelfieCheckSubmit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelfieCheckResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Selfie Check",
        "tags": [
          "users"
        ]
      }
    },
...
    "/api/v1/users/me": {
      "get": {
        "operationId": "me_api_v1_users_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Me",
        "tags": [
          "users"
        ]
      }
    },
...
  }
}